v2

+++ maps_lib.py
@@ -1,7 +1,6 @@
 import matplotlib.pyplot as plt
 from matplotlib.colors import LinearSegmentedColormap
 import pandas as pd
-from subprocess import call
 import numpy as np
 
 def n_to_one_normalizer(s,n=0):
@@ -43,12 +42,14 @@ def data_to_rgb(serie,color_maper=plt.cm.get_cmap("Blues_r"), normalizer = n_to_
 from bs4 import BeautifulSoup    
 from IPython.display import Image, display, HTML, SVG
 img_width = 400
-import os
 
-def make_map_from_svg(series_in, svg_file_path, outname, color_maper=plt.cm.get_cmap("Blues"), label = "", new_title=None):
+import os, shutil
+from subprocess import Popen, PIPE, call 
+
+def make_map_from_svg(series_in, svg_file_path, outname, color_maper=plt.cm.get_cmap("Blues"), label = "", outfolder ="img/" , new_title=None, verbose=True):
     """Makes a cloropleth map and a legend from a panda series and a blank svg map. 
     Assumes the index of the series matches the SVG classes
-    Saves the map in SVG, and in PNG if Inksscape is inkscape.
+    Saves the map in SVG, and in PNG if Inkscape is installed.
     if provided, new_title sets the title for the new SVG map
     """
     
@@ -65,17 +66,15 @@ def make_map_from_svg(series_in, svg_file_path, outname, color_maper=plt.cm.get_
        fill: {color};
        stroke:#000000;
        stroke-width:2;
-       fill-rule:evenodd;
     }}"""
 
     #Default style (for regions which are not in series_in)
     style =\
     """.default
     {
-    fill: #e0e0e0;
+    fill: #bdbdbd;
     stroke:#ffffff;
     stroke-width:2;
-    fill-rule:evenodd;
     }
     """
 
@@ -85,18 +84,10 @@ def make_map_from_svg(series_in, svg_file_path, outname, color_maper=plt.cm.get_
         style= style       + style_base.format(depname=c,color=color[c])+ "\n"   
 
 
-    #makes the legend
-    l = make_legend(100*series_in,color_maper,label,"legend_of_"+outname)
-    
-    #makes the map
-    # m= append_styles_to_map(svg_file_path,style,"map_of_"+outname,new_title)    
+    target_name = outfolder+"map_of_"+outname
 
-    target_name = "map_of_"+outname
-
-    # def append_styles_to_map(svg_file_path,style,target_name,new_title=None):
-    
-    #read input MIND UTF8
-    with open(svg_file_path, 'r',encoding='utf8') as svgfile:
+    #read input 
+    with open(svg_file_path, 'r',encoding='utf8') as svgfile: #MIND UTF8
         soup=BeautifulSoup(svgfile.read(),"xml")
 
     #names of regions to lower case without space   
@@ -104,10 +95,10 @@ def make_map_from_svg(series_in, svg_file_path, outname, color_maper=plt.cm.get_
         p["class"]=p["class"].lower().replace(" ","_")
         #Update the title (tooltip) of each region with the numerical value (ignores missing values)
         try:
-            p.title.string += "{val:.1%}".format(val=series_in[p["class"]])
+            p.title.string += "{val:.3%}".format(val=series_in[p["class"]])
         except:
             pass
-
+   
     #remove the existing style attribute (unimportant)
     del soup.svg["style"]
     
@@ -124,35 +115,70 @@ def make_map_from_svg(series_in, svg_file_path, outname, color_maper=plt.cm.get_
     with open(target_name+".svg", 'w', encoding="utf-8") as svgfile:
         svgfile.write(soup.prettify())
         
-    #inkscapes SVG to PNG    
-    call("inkscape -f {map}.svg -e {map}.png -d 150".format(map=target_name), shell=True) 
+   
+    #Link to SVG
+    display(HTML("<a target='_blank' href='"+target_name+".svg"+"'>SVG "+new_title+"</a>"))  #Linking to SVG instead of showing SVG directly works around a bug in the notebook where style-based colouring colors all the maps in the NB with a single color scale (due to CSS)
     
-    #Works around a bug in the notebook where style-based colouring colors all the maps  in the NB with a single color scale
-    # m= SVG("{map}.svg".format(map=target_name))
-    m= Image("{map}.png".format(map=target_name) ,width=img_width)  
-    display(HTML("<a target='_blank' href='"+target_name+".svg"+"'>SVG "+new_title+"</a>"))
-
-    #Attempts to downsize to a fix width and concatenate using imagemagick
-    call("convert legend_of_{outname}.png -resize {w} small_legend.png".format(outname=outname,w=img_width), shell=True )
-    call("convert map_of_{outname}.png -resize {w} small_map.png".format(outname=outname,w=img_width) , shell=True)
-    call("convert -append small_map.png small_legend.png map_and_legend_of_{outname}.png".format(outname=outname) , shell=True)
     
-    #removes temp files
-    if os.path.isfile("small_map.png"):
-        os.remove("small_map.png")
-    if os.path.isfile("small_legend.png.png"):
-        os.remove("small_legend.png.png")
+    #reports missing data        
+    if verbose:        
+        places_in_soup = [p["class"] for p in soup.findAll("path")]        
+        data_missing_in_svg = series_in[~series_in.index.isin(places_in_soup)].index.tolist()
+        data_missing_in_series = [p for p in places_in_soup if (p not in series_in.index.tolist())]
         
-    if os.path.isfile("map_and_legend_of_{outname}.png".format(outname=outname)):
-        return Image("map_and_legend_of_{outname}.png".format(outname=outname))
+        back_to_title = lambda x: x.replace("_"," ").title()
+    
+        if data_missing_in_svg:
+            print("Missing in SVG: "+"; ".join(map(back_to_title,data_missing_in_svg)))
+        if data_missing_in_series:
+            print("Missing in series: "+"; ".join(map(back_to_title,data_missing_in_series)))
+
+ 
+ 
+    if shutil.which("inkscape") is None:
+        print("cannot convert SVG to PNG. Install Inkscape to do so.")
+        could_do_png_map = False
+    else:
+        #Attempts to inkscapes SVG to PNG    
+        process=Popen("inkscape -f {map}.svg -e {map}.png -d 150".format(map=target_name, outfolder = outfolder) , shell=True, stdout=PIPE,   stderr=PIPE)
+        out, err = process.communicate()
+        errcode = process.returncode
+        if errcode:
+            could_do_png_map = False
+            print("Could not transform SVG to PNG. Error message was:\n"+err.decode())
+        else:
+            could_do_png_map = True
+
+    #makes the legend with matplotlib
+    l = make_legend(100*series_in,color_maper,label,outfolder+"legend_of_"+outname)
+    
+    if shutil.which("convert") is None:
+        print("Cannot merge map and legend. Install ImageMagick® to do so.")
+    elif could_do_png_map:
+        #Attempts to downsize to a single width and concatenate using imagemagick
+        call("convert "+outfolder+"legend_of_{outname}.png -resize {w} small_legend.png".format(outname=outname,w=img_width), shell=True )
+        call("convert "+outfolder+"map_of_{outname}.png -resize {w} small_map.png".format(outname=outname,w=img_width) , shell=True)
+        
+        merged_path = outfolder+"map_and_legend_of_{outname}.png".format(outname=outname)
+        
+        call("convert -append small_map.png small_legend.png "+merged_path, shell=True)
+        
+        #removes temp files
+        if os.path.isfile("small_map.png"):
+            os.remove("small_map.png")
+        if os.path.isfile("small_legend.png"):
+            os.remove("small_legend.png")
+    
+        if os.path.isfile(merged_path):
+            return Image(merged_path)
         
     
 import matplotlib as mpl
 
 def make_legend(serie,cmap,label="",path=None):
-    
     #todo: log flag
-        
+
+    
     fig = plt.figure(figsize=(8,3))
     ax1 = fig.add_axes([0.05, 0.80, 0.9, 0.15])
 
@@ -172,58 +198,6 @@ def make_legend(serie,cmap,label="",path=None):
     if path is not None:
         plt.savefig(path+".png",bbox_inches="tight",transparent=True)  
     plt.close(fig)    
+    
+    
     return Image(path+".png", width=img_width   )  
-    
-############################################################################# 
-#######################            FROM R             ####################### 
-#############################################################################
-import glob, sys, shutil
-from subprocess import call, Popen, PIPE
-
-def map_with_r(df_maps,map_name,**kwargs):
-    
-    df_maps.index =df_maps.index.str.lower()
-    df_maps.index.name="id"
-    df_maps.to_csv("input_data_for_map.csv",header=True)
-
-    #Writes the new R code with the arguments given to this function
-    with open("make_map.r", "wt") as fout:
-        with open("make_map.r.model", "rt") as fin:
-            for line in fin:
-                fout.write(line.format(**kwargs))
-    
-    
-    p=Popen("rscript make_map.r", stdout=PIPE);
-    print("Making the map....")
-    sys.stdout.flush()
-
-    (output, err) = p.communicate()
-    exit_code = p.wait()
-
-    if exit_code:
-        print(err,output)    
-    else:
-        print("Map done")
-
-        
-    print("There was no data for the following districts: "+", ".join(pd.read_csv("missing_admin_zones.txt",sep=" ",encoding="latin-1",squeeze=True)))    
-    
-    shutil.copy("output_map.png",map_name)
-    
-    return 
-
-############################################################################# 
-#######################         FROM GEOPANDAS        ####################### 
-############################################################################# 
-import geopandas as gpd
-def map_from_gpd(df,title="Risk to assets",col_to_plot="risk_to_assets",figname="risk_to_assets.png",cm="BuPu"):
-
-    plt.figure(figsize=(9,9))
-
-    df.plot(column=col_to_plot,colormap=cm)
-
-    plt.axis("off")
-
-    plt.title(title)
-    
-    plt.savefig(figname)
